requirements.txt
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.59.1
frozendict==2.4.6
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
lxml==6.0.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==1.23.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2025.2
pyzmq==27.0.1
requests==2.32.5
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
xlsxwriter==3.2.5
yfinance==0.2.65
zipp==3.23.0
